Run TypeInference with multiple functions that depend on each other
Summary:
To allow correctly inferring the types of variables used across
functions without being dependent on the order of the function iteration
in the IR, run TypeInference on multiple functions at once.

If a function captures another's variables or calls it, then it gets
included in the same group.

Iterate on the instructions, variables, etc across all functions
until a fixed point is reached for that group.

This way, there is no dependency between groups and the best types for
each group are inferred.

Reviewed By: neildhar

Differential Revision: D51534080

fbshipit-source-id: c54fffb37a85dfb3b339ab708f92cc9fe8566cb5