add support for --source-map
Summary:
`--source-map` specifies an input source map for a single input file.
If there are multiple input files, it is an error to use it.

We apply the source map after parsing, because we want parse errors to
be reported untranslated.

A few places had to be changed to print the translated coordinates. This
was never seriously used in Hermes, so it is mostly untested.

Fix the source map test which was incorrectly passing and add another
one.

Reviewed By: avp

Differential Revision: D51737595

fbshipit-source-id: 7128128134f5a750433e1838ceb3df71ad85c6b7