Remove unnecessary HostObject trackers in TraceInterpreter
Summary:
It turned out that these maps are holding the HostObjects and
HostFunctions created during replay, and never be released until the
runtime instance is destroyed. This changes the behavior of GC for the
host object during replay compared to tracing time as there objects can
never be collected. These maps are used only for assertions to make sure
the created IDs are unique, but they are created by the real runtime's
function, so such assumptions should be able to be made already. Hence,
simply removing these members and usages.

Reviewed By: neildhar

Differential Revision: D51874351

fbshipit-source-id: 1ca93c35b4abbc78b95e0a24419f820e16946fac