Simplify TraceInterpreter getCalls()
Summary:
With the previous change, we can simplify the logic inside getCalls()
that is using stack. The stack used to contain ObjectId or FunctionId
to get Calls vector for each case, but we just need to stack the
reference to the Call vector itself. Only for the case of
GetNativePropertyNamesRecord, we need ObjectID.

Reviewed By: neildhar

Differential Revision: D51998012

fbshipit-source-id: b9c43b5e3ec2e0d8e1f79f36cfc4e5cf41239bab