Create yarn script for prettier plugin development
Summary:
I've added a script to make development of the prettier plugin easier.

This script can be run with the following:
```
yarn build-prettier
```
It will do the following things:
1. Clone the remote fork to your devserver if necessary (inside a .gitignored directory)
1. Determine if there are any upstream commits you need to pull to ensure we dont accidentally lose any changes
1. run `yarn install`
1. copy the `dist` folder from local `hermes-parser` dir into node_modules
1. run `yarn build --no-minify`
1. copy the necessary assets to the proper directory in our source tree

NOTE: Developers will still need to ensure that they commit any prettier changes to the upstream repo and create and merge a pull request.

Reviewed By: SamChou19815

Differential Revision: D52149102

fbshipit-source-id: 5f800835957181fc0837edea8dd5d562ace6101e