shermes: support sourceMappingURL in in source comments
Summary:
Implement optional automatic detection of `//# sourceMappingURL=` magic
source comments and auto-loading of the associated source map.

The new behavior is controlled by the new CLI flag `--sm-comment` with
the following possible values:
- "off": magic comments are ignored.
- "data": magic comments are only processed if the source map is
"inline", i.e. the contents of the source map are embedded in a data
URL.
- "file": magic comments with data: or file: URLs are automatically
processed.

The default mode is "file". "data" mode can be useful for increased
security, to prevent the compiler from trying to access arbitrary file
locations.

The CLI option "--source-map" overrides this setting. If specified, it
will be used.

Add tests.

Reviewed By: avp

Differential Revision: D52149457

fbshipit-source-id: b201742d951ec6301b64861972ea0e0f913c3fe7