Don't crash in the fatal error handler on desktop
Summary:
JSI installs a fatal error handler which deliberately causes a crash.
This is intended to be handled by the crash logging of the app. However
it isn't suitable behavior when not shipping a mobile app with such
infra.

For now, keep the crash only on mobile platforms, but we likely need a
better plan for this.

Reviewed By: neildhar

Differential Revision: D52044092

fbshipit-source-id: 98e245561a2bbe820a79360ad84cddedc3fc1306