Load home object from variable during class construction
Summary:
Instead of loading the prototype property of the constructor when
instantiating a class, retrieve the saved home object from the frame.
This makes the load more efficient, and allows unused loads to be
eliminated, if the constructed object turns out to be unused or can be
promoted to the stack.

Change the IRGen for creating a new class to report an error if the
class definition has not be generated yet, since this means that the
home object would not have been populated yet, and constitutes a
compile time error.

Reviewed By: tmikov

Differential Revision: D52274399

fbshipit-source-id: 2912ddb05afbc38616fa077e243f31243e2baf50