Emit StringConcat in InstSimplify
Summary:
InstSimplify can turn `+` into `StringConcatInst` whenever both the
operands are known to be strings.

It can also simplify existing `StringConcat` by concatenating
consecutive string literals and flattening in other `StringConcat` which
are only used in that one place.

This reduces the time to run MiniReact on my M1 from 890ms to 800 ms.

Reviewed By: tmikov

Differential Revision: D52306003

fbshipit-source-id: 5ac7344593ed3f96165e5443992fab7fd5b94dc3