Add SH backend support for StringConcatInst
Summary:
StringConcatInst can immediately concatenate because it doesn't have to
do any typechecks or conversions.

There's a fast path for 2 arguments because we already have a
`StringPrototype::concat` that has some optimizations that work for that
case and it improves performance more on MiniReact.

The general case just uses StringBuilder.

Reviewed By: tmikov

Differential Revision: D52306001

fbshipit-source-id: d4bd6a311b52fa0beb835d0a6265c3f9e96bda2d