Passing num digits to represent a double into its BigInt `fromDouble` parsing
Summary:
Currently, we'll always allocate a buffer of 1024 digits in order to cast a double to a BigInt. This makes sense in the limit, as doubles can represent 1024-bit numbers.

In lots of current paths, though, we have access to the number of digits before we actually allocate this buffer. This is *really* expensive in practice. As such, let's just allocate enough space to actually represent the number we're converting rather than the largest possible representible number.

Reviewed By: mshamis85

Differential Revision: D52262602

fbshipit-source-id: a02d11f85e39496e159ed5c32a61a278a3ebe4e7