Special-casing `int32`-convertible nums in `fromDouble`
Summary:
Adds a special-case to check for numbers which are storable in an `int32`, and adds a short-circuit if so.

It looks like a lot of time is spent in the `abstractEqualityTest_RJS` method, of which the lion's share is in `BigIntPrimitive::fromDouble`

This function allocates a big buffer capable of storing 1024-bit values, since doubles can represent ints (though very few) up to that size. We then spend cycles trimming extra bytes that aren't needed until this buffer is shrunk-to-fit. This is a separate thing, and can be optimized separately and probably get big wins for `fromDouble` in general as a follow-up.

Instead, if our number fits in a 32-bit int, let's just allocate enough bits for the particular number we're setting, set the value specifically to that number, and return (rather than over-allocating and shrinking).

Reviewed By: avp

Differential Revision: D52234978

fbshipit-source-id: a097151a71b88e6221bfc5323d38c3fe26dcf0b1