RecursiveVisitor: enable optional mutation
Summary:
Visitor methods can optionally be declared to receive a pointer to the
node pointer (`Node **`) they are processing, which can then be modified
in-place. This enables very easy mutation of cases where the "parent" is
not a `NodeList`. Unfortunately that is almost never the case, because
expressions as well as statements are often included in lists.

`NodeList` requires more processing, because the removed/replaced node
might have been moved into a different list. Thus, once the visitor has
executed, we can't touch it and must take extra care to remove without
accessing it at all.

To facilitate this, carefully add two new methods to llvh::simple_ilist:
- _before_begin() returns an iterator to the sentinel.
- _erase_between(A, B) removes elements between A and B exclusive,
  without touching any of the removed elements.

While this is not very expensive, it is cost that shouldn't be incurred
by read-only visitors. Thus, it is opt in by declaring the public
constant `kEnableNodeListMutation` in the visitor class.

Force inlining of the visitChildren() methods, since they were generated
out of line by Clang, creating tons of unnecessary small functions.

Reviewed By: avp

Differential Revision: D52438214

fbshipit-source-id: 1b3410442b582f7717daea869cb184f9097cafb4