Lower fast array push into a builtin
Summary:
Instead of detecting the pattern `(x: array).push()` in IRGen, detect it
with ASTLowering and replace it with the private builtin
`?fastArrayPush`.

Private builtins is a new but pretty intuitive concepts. These are
builtins that are not recognized by FlowChecker, so they can only be
emitted internally by the compiler.

This change made it clear that the previous implementation of `push()`
had a small bug - it is supposed to return the new array length.

Reviewed By: avp

Differential Revision: D52439426

fbshipit-source-id: 657cb4252db7106f3a764222b9616b52273a65e2