Allow CheckedTypeCast to participate in type inference
Summary:
Set CheckedTypeCast as "untyped", so its output type can be narrowed by
TypeInference. I haven't been able to construct a case where this makes
a difference, so this is a theoretical benefit.

NOTE:
This opens up the possibility of using CheckTypedCast as a replacement
of ThrowIfEmpty. I have been debating internally whether that is a good
idea - I find ThrowIfEmpty easier to reason about, plus ThrowIfEmpty
confers different intent. Instead of blocking on that decision, for now
I am proceeding with keeping ThrowIfEmpty - is is easier to unify the
two instructions later than to separate them.

Reviewed By: avp

Differential Revision: D52242526

fbshipit-source-id: e88c9c0a49c1e9868ff432819e9d4fd1e24feb98