Fix infinite loop in TypeInference by only widening args
Summary:
It was possible for the type of an argument to oscillate between a wider
and narrower type due to passing the type back and forth between the
callsite and the argument.

Prevent this (temporarily) by just unioning the types of the arguments
with their type prior to `propagateArgs` instead of overwriting.

However, this does have the result of making the types too wide,
because we're defaulting the `paramTy` at `any` instead of `notype`.
This will have to be fixed along with return type widening in a
separate diff, and will have to properly handle recursive invocations.
This diff is just to fix the infinite loop bug,
and you will observe that it slightly regresses 'this' argument type
inference in tests.

Reviewed By: neildhar

Differential Revision: D52574640

fbshipit-source-id: 351936209b37abdc1d4cbd404e0d23ca81648c0d