Fix external-mem.js test
Summary:
`gc-init-heap` is defaulted to 32M, and the current logic in GCConfig
is to constrain the max heap to be at least the size of `init-heap`.
Therefore, test `external-mem.js` will not work because it won't fail
the memory allocations, because the heap size is larger than 8M.

Fix this issue by also specifying `-gc-init-heap=8M`.

Reviewed By: neildhar

Differential Revision: D52305067

fbshipit-source-id: cf60f97b73c0b038dcc0795302c2e665322f1ae2