Add CallWithArgCountInst
Summary:
Create a new IR instruction `CallWithArgCountInst`. This is done solely
to be able to emit `CallWithNewTargetLong`. In order to make
`CallWithNewTargetLong` smaller, we want to encode the number of
arguments as a register. However, the existing `BaseCallInst` does not
have the arg count as an explicit operand, so there is no register
allocated for it. Putting the argument count number literal before the
list of arguments requires changes to `BaseCallInst`'s tracking of the
operand index corresponding to `this`, or the beginning of the call
arguments. It cannot be a hardcoded enum value in `BaseCallInst`, since
we now need to allow be an arbitrary number of operands between
`NewTarget` and the beginning of the call arguments. However, the
subclasses do know exactly where the hardcoded index is for `this`, so
they can define `ThisIdx`.

Reviewed By: neildhar

Differential Revision: D52094771

fbshipit-source-id: 70972800b842ccd459c79603ff6cd769fbe61cab