Update the Github issue template
Summary:
The issue template was misleading by implying that we might be able to
support non-current versions of Hermes, which is definitely not the
case. React Native occasionally backports fixes, but that is only after
the fix has been applied to the current version.

The stable ABI will improve this, but we are not there yet.

I also attempted to address a couple of things that bugged me:
- Asking for a Hermes version is nonsensical and has been for a long
time.
- It made it seem like reporting the OS is optional.

Frankly, I am still not very happy with the template, but I hope that at
least these are changes in the right direction.

Reviewed By: neildhar

Differential Revision: D52580926

fbshipit-source-id: dfcd266fbe5a30216018781350026fb62cb887ab