Add a faster 64-bit version of isAllASCII()
Summary:
The new version uses the approach from
https://github.com/nadavrot/memset_benchmark.

The key realization is that unaligned loads are fast and that it is
faster to load parts of the same data twice than to have a conditional
branch misprediction. If we are given a 9-byte buffer, we can load 8
bytes from `buf+0` and 8 bytes from `buf+1`, without checking.

On a M1 MacBook Pro, the existing 32-bit aligned version of
isAllASCII() is about 3.5 times faster than a "naive" byte by byte
version.

This new version is 2x faster than the 32-bit version, tested both with
aligned and unligned buffers.

Differential Revision: D52649631