Fix NumberFormat SetNumberFormatDigitOptions logic (#1246)
Summary:
Pull Request resolved: https://github.com/facebook/hermes/pull/1246

The old Android logic was the 2020 version of the [spec](https://402.ecma-international.org/7.0/index.html#sec-setnfdigitoptions).

The old logic passes `mnfdDefault` as the fallback to
`DefaultNumberOption()` at step 12b. This causes the `mnfd` passed to
step 12d to be whatever the default for 'currency' style is. And since
`maximumFractionDigits` is set to 0 and 0 is less than `mnfd = 2`, a
RangeError was thrown.

This problem was already fixed by the 2021 version of the spec.

`PlatformIntlApple.mm` already implements the newer spec, so that's why
it's an Android-only issue.

Since the future plan is to move Intl to use ICU4C on Android, that
means much of the Java code becomes obsolete. Still, it's relatively
easy to just update the logic to match Apple platform [spec](https://402.ecma-international.org/8.0/index.html#sec-setnfdigitoptions), so I did.

In the process, I added tests to catch this bug. The tests are probably
the more valuable part as we implement on top of ICU4C.

Reviewed By: neildhar

Differential Revision: D52641902

fbshipit-source-id: d85a415c3440b2fe81c899d8ad0fd3c6eefc8031