Mirror ABI types and helpers in the sandbox
Summary:
The layout of data structures in the sandbox is different than that on
the host. In order to make it more convenient to consume and create
these data structures, create a full mirror of all the `HermesABI*`
types such that their layout matches what it would have in the
sandbox.

When passing/returning a value from the sandbox, it just needs to be
cast to one of these types to access its fields.

Reviewed By: avp

Differential Revision: D52558676

fbshipit-source-id: 2e88e294c7c827959e304381cc135d354bc0dc51