Implement utilities for pointers and allocations in sandbox memory
Summary:
Implement some utilities to make it easier to interact with pointers
into the sandbox. Since pointers are just `uint32`s, they convey no
type information and are annoying to dereference.

This diff defines a `Ptr` template, and subclasses, that allow you to
treat the pointer like a reference to the underlying object. Note that
this is safer than directly holding a pointer to the underlying memory,
because the sandbox heap can move when it grows.

The subclasses make it convenient to allocate data on the stack/heap
for the duration of the implementation of JSI functions.

Reviewed By: avp

Differential Revision: D52558679