Create and destroy the sandboxed runtime
Summary:
Instantiate a runtime object through the ABI by retrieving the vtable
and calling `make_hermes_runtime`. The object is owned by the JSI
runtime and is released on destruction.

Read out the vtable fields as correctly typed function pointers to
populate the mirrored vtable `vt_`. These will be the primary mechanism
for calling into the sandbox.

Reviewed By: avp

Differential Revision: D52558677

fbshipit-source-id: 770d97291a6985d2b62f747b0c3b804caef5b56c