FlowChecker: narrow type of assignment expression
Summary:
This changes addresses two issues.

The first is a bug: the result type of the assignment expression should
never be wider than the type of the LHS.

The second is an improvement. Occasionally, the type of the RHS is
narrower than LHS. In those cases we can use the narrower type as the
result, since this is the value we are actually storing.

This helps cases like this one:
```
        let u: string|number, n: number;
        n = u = 5;
```

Previously this would have produced an error.

Note that Flow also accepts this, but it also supports a more "advanced"
version, which we don't:
```
        let u: string|number, n: number;
        u = 5; // Flow knows u has type number now.
        n = u; // Flow allows this, but we don't yet.
```

Differential Revision: D52791641

fbshipit-source-id: 47652c969a216729ee6481e11041d085cf6b7704