Add a new IR type Uninit
Summary:
`Uninit` is used to "poison" a typed variable, after it has been
declared, but before it has been initialized.

In the following code:

    {
      /* 1 */
      let x: Foo; /* 2 */
      x = new Foo(); /* 3 */
    }

`x` is initialized to `Empty` in position 1, then initialized to
`Uninit` in position 2, before it gets its actual value in 3.

We are not using the new type yet. This is just adding it.

Differential Revision: D52296535

fbshipit-source-id: b48ec2e9391a0e0c93b4c8db4440fe81e91d2c1c