Add a new IR type Uninit
Summary:
`Uninit` is used to "poison" a typed variable, after it has been
declared, but before it has been initialized.

In the following code:

    {
      /* 1 */
      let x: Foo; /* 2 */
      x = new Foo(); /* 3 */
    }

`x` is initialized to `Empty` in position 1, then initialized to
`Uninit` in position 2, before it gets its actual value in 3.

We are not using the new type yet. This is just adding it.

Reviewed By: avp

Differential Revision: D52296535

fbshipit-source-id: 1d44d7b583d5ee5ac5785d5516cf2912e03ffe8f