Add packages concept to flow-bundler
Summary:
Add the concept of packages to the `flow-bundler` so we can better match existing project structures and update the MiniReact demo to use this. This allows us to reuse existing code that relies on packages without having to rewrite their imports, e.g. `import * as React from 'react'` instead of `import * as React from '../../../react'`. It also makes projects easier to manage as they grow in scale as you no longer need to update as many relative paths as code is moved around. This then allows our demo code to look more like code that developers write today.

This was a fairly simple change for the flow-bundler, i just changed the path resolution to resolve paths into the root folders "packages" directory if the path is not absolute or relative. I went with "packages" as the directory name and not "node_modules" since i don't expect us to directly be able to npm install packages without modifying them yet and "node_modules" are excluded by source control atm so this would require larger infra changes.

I also added support for root relative paths, e.g. `@/App.js`, since this is used within a couple of the UI packages im looking at. I also set it up to work within packages (so it will use the package root instead of the project root).

Reviewed By: avp

Differential Revision: D52714156

fbshipit-source-id: 612db480ac3002dc3dac2a4f564482d726512bbd