Remove regex breakpoints
Summary:
Remove handling of URL regexes. This functionality is not not used in
our current configurations.

The only reference to regex breakpoints in Chrome DevTools appears
to be for debugging a NodeJS target on Windows, when running
scripts from local filepaths (it's used to work around the case
insensitivity of windows paths).

Our use cases only make use of setting breakpoints by URL or by
script ID.

This change makes the subsequent stack simpler, and removes use of
`std::regex`, which is considered problematic.

Reviewed By: dannysu

Differential Revision: D51674836

fbshipit-source-id: e547738cbe5b0f571f87d8a206002ff2852231f2