Fixed symbol conflicts when using identical function and method names in an ES6 class (#1257)
Summary:
This change fixes an issue that could occur when calling a function from a method that has the same name. Consider this example:
```js
function myMethod() {
  console.log('Hello from function');
}

class MyClass {
  myMethod() {
    myMethod();
  }
}
```
In this scenario, `Hello from function` would not be printed when `instance.myMethod()` is called as the method would be called recursively instead of the function that is available in the outer scope.

## Changes
- Added test showcasing the issue
- Always create an anonymous function for methods (maybe there is a way to make the identifier private instead?), so that the method body cannot recursively refers to itself by its name.
- Insert `method.name` at the JS level when the method is about to be registered into the class.

Pull Request resolved: https://github.com/facebook/hermes/pull/1257

Test Plan:
All tests passes.
<!--
  Demonstrate the code is solid.
  Example: The exact commands you ran and their output,
  screenshots / videos if the pull request changes the user interface.
-->

Reviewed By: avp

Differential Revision: D52824378

Pulled By: tmikov

fbshipit-source-id: e224732fea670680250dee50e964e8e91c1ab1f0