IRGen: cast all expressions to their Flow types
Summary:
If the IR type from a compiled expression doesn't match the type the
Flow checker assigned to it, emit a CheckedCastInst.

This mostly happens automatically by wrapping `genExpression()`. The
update and compound assignment expressions also had to be changed.

Naturally this results in a lot of extra casts to remind us that IRGen
is not finished. The casts will automatically go away as we add typed
code generation everywhere in IRGen.

Reviewed By: avp

Differential Revision: D52771166

fbshipit-source-id: beb120a6a7d8330afba994bbd8d59b00a9620af8