Fix Proxy defineProperty trap
Summary:
Original Author: fumi@meta.com
Original Git: 7d93640ee9000363fce313188f876ec3d4e8a4f1
Original Reviewed By: fbmal7
Original Revision: D51327887

When a data property of a proxied object is assigned a new value, the defineProperty trap of Proxy is triggered with a descriptor with configurable: false, enumerable: false, and writable: false. However, according to the ECMAScript spec (see item 2.d.iii), the descriptor object must have no field other than value. This is problematic if the Proxy attempts to call Reflect.defineProperty() using the passed descriptor, as it unexpectedly "freezes" the property.

This is happening because `objectFromPropertyDescriptor` always construct an object with the fields set for writeable, enumerable, and configurable as `ComputedPropertyDescriptor` doesn't have notion of whether these fields exist or not.

Fixing this by changing to pass `DefinePropertyFlags` instead to `objectFromPropertyDescriptor` so that it can set the fields based on the existence of the fields.

Reviewed By: tmikov

Differential Revision: D52854467

fbshipit-source-id: 1e48faa2ce291d45572045562400343fbee26e8b