Initial support for Hooks Syntax
Summary:
I've added initial support for hooks syntax to the parser. This diff contains only the parser code itself. The next diff in the stack includes the hermes-parser js changes.

NOTE: This is my first contribution to the hermes parser implementation, so please take a close look and let me know if I'm missing anything!

Reviewed By: avp

Differential Revision: D51875900

fbshipit-source-id: 75650b7fa070916a54b882728cfb9912812c4e6d