Added support for providing bytecode generation option to BytecodeProviderFromSrc (#1262)
Summary:
This change adds support for providing the bytecode generation options
to the `BytecodeProviderFromSrc` API. This makes it possible to strip
the function names and debug information when using the high level
`BytecodeProviderFromSrc` API without having to re-write what
`BytecodeProviderFromSrc` or `CompilerDriver` are already doing. For
context, we currently statically embed Hermes into one of our own CLI
binary that is able to do a lot of different compile related tasks for
us, as such we are not using the Hermes CLI API. We still would like to
be able to compile modules as byte code, and the only thing missing
from `BytecodeProviderFromSrc` to be at parity with the
`CompilerDriver` CLI option that does bytecode generation was the
ability to configure the bytecode generation phase.

Pull Request resolved: https://github.com/facebook/hermes/pull/1262

Test Plan:
<!--
  Demonstrate the code is solid.
  Example: The exact commands you ran and their output,
  screenshots / videos if the pull request changes the user interface.
-->

Tests pass

Reviewed By: tmikov

Differential Revision: D52918720

Pulled By: neildhar

fbshipit-source-id: 0836a8f67a8c3c8783f346b9ee625cba3d207b6e