Implement a helper for function type signatures
Summary:
Original Author: neildhar@meta.com
Original Git: 4cd071292b22a997c9f9fa6281b929d2f4b003d9
Original Reviewed By: avp
Original Revision: D52558681

Function pointers in the sandbox are just indices into a function
pointer table. The table also includes a type signature encoded as a
`wasm_rt_func_type_t`, which is used to verify that the function being
retrieved has the expected type (since it would be unsafe to call it if
the types do not match).

So whenever we add a function to the table, or we try to retrieve a
function, it is necessary to construct its corresponding
`wasm_rt_func_type_t` to populate/check the stored type. This results
in a lot of boilerplate, and is error prone since we need to manually
ensure that the constructed `wasm_rt_func_type_t` matches the actual
signature of the function.

This diff defines a template that performs the conversion from a
function signature to a `wasm_rt_func_type_t` automatically, by
decomposing the function type and constructing the right call to
`wasm2c_hermes_get_func_type`.

Reviewed By: neildhar

Differential Revision: D52854496

fbshipit-source-id: 54b1bd4424ae9aaf3e2de6d35c6d7a8a147acaf5