Implement utilities for pointers and allocations in sandbox memory
Summary:
Original Author: neildhar@meta.com
Original Git: c924bff541a535468d9de42c21fd6318460d1acd
Original Reviewed By: avp
Original Revision: D52558679

Implement some utilities to make it easier to interact with pointers
into the sandbox. Since pointers are just `uint32`s, they convey no
type information and are annoying to dereference.

This diff defines a `Ptr` template, and subclasses, that allow you to
treat the pointer like a reference to the underlying object. Note that
this is safer than directly holding a pointer to the underlying memory,
because the sandbox heap can move when it grows.

The subclasses make it convenient to allocate data on the stack/heap
for the duration of the implementation of JSI functions.

Reviewed By: neildhar

Differential Revision: D52854377

fbshipit-source-id: 9e847e82e78e6f8e3b43e6add32e47b43c9f1229