Add asyncTriggerTimeout to sandbox
Summary:
Most sandbox users would also benefit from having the ability to
interrupt execution of untrusted JS when it is taking too long. It
would be slightly more complicated (but still possible) to implement
async timeouts in the way that they are implemented in Hermes, since
that would involve reaching into sandbox memory from another thread and
writing to it.

To avoid needing any special synchronisation with the sandbox and
exposing the internal flag, this diff instead has the runtime call a
host provided callback on every `AsyncBreakCheck` to check if a timeout
was requested. Empirically, this call has relatively low performance
overhead, and under LTO, it may even be inlined since the target of the
call is known.

allow-large-files

Reviewed By: avp

Differential Revision: D52947241

fbshipit-source-id: 1e1d777754af7c0cc378aa0df3a1f4cd84b23a91