Enable AsyncBreakCheckInEval by default
Summary:
We currently default `AsyncBreakCheckInEval` to false, which means that
source evaluated through `eval` and `evaluateJavaScript` cannot be
interrupted. This leads to some odd practical problems:
1. The new `asyncTriggerTimeout` method only works if a `RuntimeConfig`
is correctly set.
2. Code evaluated before a call to `watchTimeLimit` will not be
instrumented, but code evaluated after it will, since calling
`watchTimeLimit` enables async break checks. So if code evaluated after
`watchTimeLimit` is called then calls into a function evaluated before
it, the timeout may fail to trigger. It also seems that code generated
by `eval` will not be instrumented, even when `watchTimeLimit` is
called, leading to pretty confusing behaviour.
3. When the debugger is enabled, we set the debug compile flag to true,
which also results in the emission of async breaks. So an async break
may correctly trigger in dev mode, but not in release mode.

Switching it to be true by default results in simpler and more
predictable behaviour, while keeping the option to disable it for users
who are absolutely sure it is not needed.

Reviewed By: avp

Differential Revision: D52947242

fbshipit-source-id: 9901eee3a121f11cf13e5bc530a523cf7c3bca0b