Fix failing test
Summary:
The CDP Agent impl was being kept alive too long by queued tasks that held a shared pointer to it, causing the shutdown task for the domain handlers to be queued after the task runner had disappeared.

Remove the shared pointer so the CDP Agent impl can shut down during destruction of the CDP Agent.

This change is a blatant rip-off of dannysu's D52937432. I'm happy to have it merged instead of this diff if that works.

Reviewed By: fbmal7

Differential Revision: D52978446

fbshipit-source-id: 3a1ab66db217133a8abd26be80b72d18c0043f46