Stable bytecode encoding for qNaN
Summary:
The default bit encoding of quiet NaN is platform dependent. For
example, on x86_64 the sign bit is 1, while on arm64 it is 0. Both
encodings are "understood" by the hardware, but which one is generated
may differ.

Make sure we always emit the x86_64 bit pattern emit serializing
bytecode.

This was tested with the following input:
```
let x = 0/0;
print(x);
```

Reviewed By: avp

Differential Revision: D53098568