Have AsyncDebuggerAPI manage DebuggerAPI's isDebuggerAttached state
Summary:
Hermes DebuggerAPI has a `setIsDebuggerAttached()` function whose
purpose is so that in JavaScript one could query whether debugger is
attached via `DebuggerInternal.isDebuggerAttached`.

What it means for debugger to be attached is left up to the caller. It
is not correlated to whether there is an EventObserver.

For AsyncDebuggerAPI, there could be multiple debug clients that add
DebuggerEventCallback. It doesn't quite make sense for individual debug
clients to be calling `setIsDebuggerAttached()` and potentially run
into conflict with each other. Therefore, adding code to manage that
state centrally in AsyncDebuggerAPI.

The meaning of whether "debugger is attached" is thus "whether any
clients registered DebuggerEventCallback".

Reviewed By: mattbfb

Differential Revision: D52887165

fbshipit-source-id: 5f407833efa704e900f4bab6c639949511b21d11