Add initial tests for DebuggerDomainAgent
Summary:
Migrated some Debugger domain tests from ConnectionTests related to
enable/disable.

NOTE: In the `CDPHandler` implementation we seem to allow
`Debugger.enable` to be received when the domain is already enabled. I
changed this such that an error will be sent. This means there are
slight differences in the `TestScriptsOnEnable` and
`TestScriptsOrdering` tests.

Reviewed By: mattbfb

Differential Revision: D52803852

fbshipit-source-id: 98f73c1ddf1034b489e3f8bb812b0fbdc8d82e8f