Preserve breakpoints across instances
Summary:
Allow state to be exported from a CDP Handler instance and imported
into another. The criteria for when state should be persisted are up
to the integrator (e.g. if they know the same app is being reloaded
in a new runtime).

The state object is setup to contain arbitrary state we want to
persist. Exporting the state (`getState`) copies the data to be
persisted (breakpoint descriptions, currently) into the state object.

Importing the state (via the CDPHandler constructor) loads the relevant
data from the state into the new CDP handler. Currently, this involves
creating a breakpoint for each loaded description, and adjusting the
breakpoint ID counter.

Reviewed By: dannysu

Differential Revision: D50204914

fbshipit-source-id: f95d87d9e648110aeb41824f8a7c67ed89f0933d