Fix end-to-end Android tests (#1284)
Summary:

Intl tests in CircleCI are broken because using a newer version of the
NDK with an older version of fbjni appears to cause problems. Bump the
version of fbjni.

While here, skip a test that I observed is failing on my computer, possibly
because of a change in the default handling for some localisation
functionality in different Android versions.

Differential Revision: D53153403