Fix end-to-end Android tests (#1284)
Summary:
Pull Request resolved: https://github.com/facebook/hermes/pull/1284

Intl tests in CircleCI are broken because using a newer version of the
NDK with an older version of fbjni appears to cause problems. Bump the
version of fbjni.

While here, skip a test that I observed is failing on my computer, possibly
because of a change in the default handling for some localisation
functionality in different Android versions.

Reviewed By: lavenzg

Differential Revision: D53153403

fbshipit-source-id: dc88df5d83c063c4ca2f25f246298286fdd4ae51