Simplify importing host compiler
Summary:
Simplify the logic to import a host compiler so that we can do it even
when not cross-compiling. This allows us to set up a host build that
imports the compiler from a different host build, so that we can use
different configurations for the compiler and VM builds.

Reviewed By: tmikov

Differential Revision: D53189094

fbshipit-source-id: 169226e0ea8e54bdf44fad3227a066ed1f8fbdd0